pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ppo"
version = "0.1.0"
description = "PPO - Pluggable Portfolio Optimizer: Factor evaluation service with CLI, API, and MCP interface"
readme = "ffo/README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    # CLI
    "click>=8.0",
    "rich>=13.0",
    # Config
    "pyyaml>=6.0",
    # Server process management
    "psutil>=5.9",
    # HTTP client
    "requests>=2.28",
    # MCP server
    "fastmcp>=2.0",
    # Flask server (backend)
    "flask>=2.3",
    "flask-cors>=4.0",
    "gunicorn>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.scripts]
ppo = "ffo.cli.main:cli"

[tool.setuptools.packages.find]
# Include ffo package and all sub-packages
include = ["ffo*"]

[tool.setuptools.package-data]
ffo = [
    "config/ffo.yaml",
    "templates/**/*",
    "examples/**/*",
    "demos/**/*",
]
